明細CSV取込：Codexレビュー4回目修正

+++ backend/src/main/java/com/example/minimal/common/constants/CategoryId.java
@@ -4,9 +4,9 @@ public final class CategoryId {
 	private CategoryId() {
 	}
 
-	public static final Long UNKNOWN = 0L;
-	public static final Long FOOD = 1L;
-	public static final Long ELECTRONICS = 2L;
-	public static final Long CLOTHING = 3L;
+	public static final Integer UNKNOWN = 0;
+	public static final Integer FOOD = 1;
+	public static final Integer ELECTRONICS = 2;
+	public static final Integer CLOTHING = 3;
 
 }
